ai/requirements.txt
jupyterlab>=2.0.1
jupyter-http-over-ws>=0.0.8
matplotlib>=3.2.1
tensorflow-gpu>=2.2.0rc2
tensorboard>=2.1.0rc0
graph-nets>=1.1.0
dm-sonnet>=2.0.0
tensorflow-probability>=0.9.0
